Finished commenting all the code I wrote.

+++ lib/Organization/add_beneficiary_screen.dart
@@ -7,7 +7,7 @@ import 'package:flutter/services.dart';
 import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
 import 'organization_dashboard.dart';
 import 'package:get/get.dart';
-
+// Add campaign form
 class AddBeneficiaryForm extends StatefulWidget {
   static const id = 'beneficiary_form_screen';
   AddBeneficiaryForm({Key? key}) : super(key: key);
@@ -35,7 +35,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
   var category = [];
   int beneficiaryTimeLimit=0;
 
-
+// Get category
   _getCampaign() async {
     var ret = await firestore
         .collection('CharityCategories')
@@ -46,7 +46,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
 
     setState(() {});
   }
-
+// Get admin time limit from firebase
   _getTimeLimit() async {
     var ret = await firestore.collection('AdminRestrictions').where('id',isEqualTo: 'CharityDurationLimits').get();
 
@@ -66,7 +66,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
     loggedInUser = _auth.currentUser;
   }
 
-
+//Add beneficiary to firebase
   Future<void> addBeneficiary(String category, String biography, double goalAmount,
       String name, String endDateController) async {
     try {
@@ -89,7 +89,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
     }
   }
 
-
+// UI for for beneficiary
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -115,6 +115,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                         const SizedBox(
                           height: 15.0,
                         ),
+                        // Add beneficiary from UI
                          Center(
                           child: Text(
                             'add_beneficiary'.tr,
@@ -128,10 +129,12 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                           padding: EdgeInsets.symmetric(
                               vertical: 8.0, horizontal: 25.0),
                           child: Text(
+                            // Required field
                             '* - required_fields'.tr,
                             style: TextStyle(color: Colors.red),
                           ),
                         ),
+                        // Error message please enter nickname
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
@@ -147,6 +150,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 return null;
                               }
                             },
+                            // UI for nickname enter field
                             textAlign: TextAlign.center,
                             decoration: InputDecoration(
                                 label: Center(
@@ -172,6 +176,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 )),
                           ),
                         ),
+                        // Error message please enter bio
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
@@ -186,6 +191,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 return null;
                               }
                             },
+                            //biography field ui
                             textAlign: TextAlign.center,
                             decoration: InputDecoration(
                                 label: Center(
@@ -211,6 +217,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 )),
                           ),
                         ),
+                        //Please enter goal amount error message or enter valid goal amount
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
@@ -226,6 +233,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 return null;
                               }
                             },
+                            //UI for goal amount
                             textAlign: TextAlign.center,
                             decoration: InputDecoration(
                                 label: Center(
@@ -251,6 +259,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 )),
                           ),
                         ),
+                        //Please valid enter end date error
                         Padding(
                             padding: const EdgeInsets.all(8.0),
                             child: TextFormField(
@@ -300,7 +309,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                 endDateController.text =
                                     date.toString().substring(0, 10);
                               },)),
-
+                        // Category field UI
                         Padding(
                             padding: const EdgeInsets.all(8.0),
                             child: DropdownButtonFormField <String>(
@@ -343,6 +352,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                   ? 'please_fill_in_the_category'.tr : null,
                             )
                         ),
+                        //Submit button and navigate to the dashboard
                         Padding(
                           padding: const EdgeInsets.symmetric(
                               vertical: 25.0, horizontal: 5.0),
@@ -364,7 +374,7 @@ class _AddBeneficiaryFormState extends State<AddBeneficiaryForm> {
                                   });
                                   addBeneficiary(categoryController.text,
                                       biographyController.text,
-                                      int.parse(goalAmountController.text).toDouble(),
+                                      double.parse(goalAmountController.text).toDouble(),
                                       nameController.text,
                                       endDateController.text);
                                   Navigator.of(context).popUntil(ModalRoute.withName(OrganizationDashboard.id));
